+++ dwds/lib/src/debugging/inspector.dart
@@ -86,8 +86,10 @@ class AppInspector implements AppInspectorInterface {
   /// Flutter widget inspector library.
   static const _flutterWidgetInspectorLibraryUri =
       'package:flutter/src/widgets/widget_inspector.dart';
-  LibraryRef? flutterWidgetInspectorLibrary;
-  late bool isFlutterApp;
+  late LibraryRef? flutterWidgetInspectorLibrary;
+
+  /// Regex used to extract a stack trace line from the exception description.
+  static final stackTraceLineRegex = RegExp(r'^\s*at\s.*$', multiLine: true);
 
   AppInspector._(
     this._appConnection,
@@ -123,9 +125,8 @@ class AppInspector implements AppInspectorInterface {
 
     flutterWidgetInspectorLibrary = libraries.firstWhereOrNull(
         (lib) => lib.uri == _flutterWidgetInspectorLibraryUri);
-    isFlutterApp = flutterWidgetInspectorLibrary != null;
 
-    final appKind = isFlutterApp ? "Flutter" : "dart";
+    final appKind = flutterWidgetInspectorLibrary != null ? "Flutter" : "dart";
     _logger.finest('Debugging $appKind app');
 
     isolate.extensionRPCs?.addAll(await _getExtensionRpcs());
@@ -624,8 +625,17 @@ class AppInspector implements AppInspectorInterface {
     if (mappedStack == null || mappedStack.isEmpty) {
       return description;
     }
-    var message = exceptionMessageRegex.firstMatch(description)?.group(0);
-    message = (message != null) ? '$message\n' : '';
+    final message = _allLinesBeforeStackTrace(description);
     return '$message$mappedStack';
+  }
+
+  String _allLinesBeforeStackTrace(String description) {
+    var message = '';
+    for (final match in exceptionMessageRegex.allMatches(description)) {
+      final isStackTraceLine = stackTraceLineRegex.hasMatch(match[0] ?? '');
+      if (isStackTraceLine) break;
+      message += '${match[0]}\n';
+    }
+    return message;
   }
 }