+++ dwds/lib/src/services/chrome_proxy_service.dart
@@ -209,7 +209,7 @@ class ChromeProxyService implements VmServiceInterface {
       // cards etc.
       // Pre-warming the cache while DevTools is still loading helps
       // Flutter Inspector start faster.
-      if (inspector.isFlutterApp) {
+      if (inspector.flutterWidgetInspectorLibrary != null) {
         _logger.finest('Warming up expression compiler worker');
         unawaited(evaluate(inspector.isolateRef.id!,
             '${inspector.flutterWidgetInspectorLibrary!.uri}', 'true'));