+++ dwds/CHANGELOG.md
@@ -6,6 +6,8 @@
   internally or externally.
 - Fix null cast error on expression evaluations after dwds fails to find class
   metadata.
+- Include the entire exception description up to the stacktrace in
+  `mapExceptionStackTrace`.
 - Pre-warm expression compiler cache to speed up Flutter Inspector loading.
 
 ## 16.0.1